requirements.txt
fastapi[standard]
sqlalchemy
python-jose
passlib[bcrypt]
bcrypt==4.0.1